+++ src/main/java/org/geowe/client/local/messages/UIMessages.java
@@ -822,4 +822,8 @@ public interface UIMessages extends Messages {
 	String projectTotalVectorLayer();
 	
 	String projectDescription();
+	
+	String projectVersion();
+	
+	String projectDate();
 }